matrix_converter/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
psycopg==3.1.10
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
yarl==1.9.4